Update method for getting player with id Update Rank to long data type

+++ src/Redisboard.NET.DemoAPI/Models/Player.cs
@@ -6,7 +6,7 @@ public class Player : ILeaderboardEntity
 {
     public string Id { get; set; }
     
-    public int Rank { get; set; }
+    public long Rank { get; set; }
     
     public double Score { get; set; }
